requirements.txt
Django>=4.0
djangorestframework
requests
beautifulsoup4
django-cors-headers
gunicorn
